pyproject.toml
[project]
name = "openid-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Sani", email = "nitori@chireiden.net" }
]
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "argon2-cffi>=25.1.0",
    "cryptography>=46.0.5",
    "email-validator>=2.3.0",
    "flask>=3.1.3",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "libpass>=1.9.3",
    "loguru>=0.7.3",
    "platformdirs>=4.9.2",
    "psycopg>=3.3.3",
    "pyjwt>=2.11.0",
    "pyotp>=2.9.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "waitress>=3.0.2",
]

[project.scripts]
openid-server = "openid_server:main"

[build-system]
requires = ["uv_build>=0.9.28,<0.10.0"]
build-backend = "uv_build"
